requirements.txt
pyserial==3.4
pyudev==0.21.0
requests==2.20.1
rshell==0.0.16
esptool==2.5.1
paho-mqtt==1.4.0
mpfshell==0.9.0
